LayoutTests/accessibility/hidden-legend-expected.txt

Pick your favourite colour:
 red  yellow
Pick your favourite colour: red  yellow
 red  yellow
This tests that the legend is still used as the title UI element even when off-screen (but not when display:none is used)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fieldset1.titleUIElement().isEqual(fieldset1.childAtIndex(0)) is true
PASS fieldset2.titleUIElement().isEqual(fieldset2.childAtIndex(0)) is true
PASS fieldset3.titleUIElement() == null || !fieldset3.titleUIElement().isValid is true
PASS successfullyParsed is true

TEST COMPLETE
